requirements.txt
PyMuPDF==1.25.1
langchain==0.3.12
langchain-google-genai==2.0.7
python-dotenv==1.0.1
flask==3.1.0